% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/china.map.R
\name{china.map}
\alias{china.map}
\title{China map based ggplot2 version 1}
\usage{
china.map(
  data,
  fill.variable = "value",
  limits = c(4000, 19000),
  midpoint = 12000,
  breaks = c(5000, 8000, 11000, 14000, 17000)
)
}
\arguments{
\item{data}{a data frame containing values to fill the provinces}

\item{fill.variable}{a numeric vector}

\item{limits}{the range of fill variable}

\item{midpoint}{the midpoint of fill variable}

\item{breaks}{the breaks of fill variable}
}
\value{

}
\description{
China map based ggplot2 version 1
}
\examples{
data <- openxlsx::read.xlsx("provincedata.xlsx", sheet = 1, rowNames = F, colNames = T)
library(showtext)
showtext_auto(enable = TRUE)
font_add('Songti', 'Songti.ttc','STKaiti','STXihei')
china.map(data = data, fill.variable = "value",
limits = c(4000, 19000),
midpoint = 12000,
breaks = c(5000, 8000, 11000, 14000, 17000))
}
\author{
ZhonghuiGai
}
