% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcnmap.R
\name{ggcnmap}
\alias{ggcnmap}
\title{China map based ggplot2 version 2}
\usage{
ggcnmap(
  data,
  fill.variable = "Value",
  save = FALSE,
  low = "#3c9eff",
  high = "#ff445d",
  mid = "#f8d248",
  fill = 2,
  limits = c(4000, 19000),
  midpoint = 12000,
  breaks = c(5000, 8000, 11000, 14000, 17000)
)
}
\arguments{
\item{data}{a data frame containing values to fill the provinces}

\item{fill.variable}{a numeric vector}

\item{low}{low color}

\item{high}{high color}

\item{mid}{the color of midpoint}

\item{limits}{the range of fill variable}

\item{midpoint}{the midpoint of fill variable}

\item{breaks}{the breaks of fill variable}
}
\value{
a ggplot
}
\description{
China map based ggplot2 version 2
}
\examples{
data <- ggcnmap::province
library(showtext)
showtext_auto(enable = TRUE)
font_add('Songti', 'Songti.ttc','STKaiti','STXihei')
ggcnmap(data = data, fill.variable = "value", fill = 2,
        limits = c(40, 600),
        midpoint = 5,
        breaks = c(50, 80, 110, 140, 600))
}
\author{
ZhonghuiGai
}
